complited task

+++ src/services/contacts.js
@@ -1,20 +1,24 @@
-import { Contact } from '../models/contactModel.js';
+import { Contact } from '../db/models/contactModel.js';
 import { SORT_ORDER } from '../constants/index.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
+
 export const getAllContacts = async ({
   page = 1,
   perPage = 10,
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
-
+  const filter = { userId };
   const contactsQuery = Contact.find();
 
-  const contactsCount = await Contact.find()
-    .merge(contactsQuery)
-    .countDocuments();
+  // const contactsCount = await Contact.find()
+  //   .merge(contactsQuery)
+  //   .countDocuments();
+
+  const contactsCount = await Contact.countDocuments(filter);
 
   const contacts = await contactsQuery
     .skip(skip)
@@ -30,16 +34,16 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (contactId) => {
-  return await Contact.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+  return await Contact.findById({ _id: contactId, userId });
 };
 export const createContact = async (data) => {
   const contact = new Contact(data);
   return await contact.save();
 };
-export const updateContactById = async (contactId, updateData) => {
+export const updateContactById = async (contactId, userId, updateData) => {
   const updatedContact = await Contact.findByIdAndUpdate(
-    contactId,
+    { _id: contactId, userId },
     updateData,
     {
       new: true,
@@ -48,7 +52,7 @@ export const updateContactById = async (contactId, updateData) => {
   );
   return updatedContact;
 };
-export const deleteContactById = async (contactId) => {
-  const result = await Contact.findByIdAndDelete(contactId);
+export const deleteContactById = async (contactId, userId) => {
+  const result = await Contact.findByIdAndDelete({ _id: contactId, userId });
   return result;
 };
